requirements.txt
bidict==0.22.0
black==22.8.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
flake8==5.0.4
idna==3.3
mccabe==0.7.0
mypy-extensions==0.4.3
paho-mqtt==1.6.1
pathspec==0.10.1
platformdirs==2.5.2
pycodestyle==2.9.1
pydantic==1.10.2
pyflakes==2.5.0
python-dotenv==0.21.0
python-engineio==4.3.4
python-mpv==1.0.1
python-socketio==5.7.1
python-vlc==3.0.16120
requests==2.28.1
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
websocket-client==1.4.1
